+++ Game/AI/ScriptedNodes/Utils/SCR_AIGetPosForwardToTarget.c
@@ -59,7 +59,7 @@ class SCR_AIGetPosForwardToTarget: AITaskScripted
 		{
 			int distance;
 			GetVariableIn(PORT_DISTANCE, distance);
-			m_Distance = (float)distance;
+			m_Distance = distance;
 		}
 		else if (GetVariableType(true, PORT_DISTANCE) == float)
 		{
